traceflow and tcp.hpp together form something very similar to what tcpflow does, but in much cleaner code.
- Haven't memory-leak tested it yet but I have run it through basic functionality tests.
- No unit tests for it.

+++ test/unit-collator.cpp
@@ -27,7 +27,7 @@ struct hc
     int i;
 };
 
-void handler(void* context, uint32_t length, void* data)
+void handler(void* context, uint64_t length, void* data)
 {
     struct hc* ctxt = (struct hc*)context;
     while (len(g_i) == 0)
